+++ packages/espressocash_app/lib/data/db/db.dart
@@ -36,6 +36,7 @@ const _tables = [
   OnRampOrderRows,
   OffRampOrderRows,
   OutgoingDlnPaymentRows,
+  TransactionRequestRows,
 ];
 
 @lazySingleton
@@ -101,10 +102,7 @@ class MyDatabase extends _$MyDatabase {
             await m.addColumn(offRampOrderRows, offRampOrderRows.feeToken);
           }
           if (from < 49) {
-            await m.addColumn(
-              paymentRequestRows,
-              paymentRequestRows.resolvedAt,
-            );
+            await m.createTable(transactionRequestRows);
           }
         },
       );
@@ -207,4 +205,20 @@ class TransactionRows extends Table {
 
   @override
   Set<Column<Object>> get primaryKey => {id};
+}
+
+class TransactionRequestRows extends Table with AmountMixin, EntityMixin {
+  const TransactionRequestRows();
+
+  TextColumn get label => text()();
+  TextColumn get transaction => text()();
+  Int64Column get slot => int64()();
+  TextColumn get status => textEnum<TRStatusDto>()();
+}
+
+enum TRStatusDto {
+  created,
+  sent,
+  success,
+  failure,
 }