+++ packages/espressocash_app/lib/features/feature_flags/services/feature_flags_manager.dart
@@ -25,6 +25,9 @@ class FeatureFlagsManager {
 
   bool isIncomingDlnEnabled() =>
       _remoteConfig.getBool(FeatureFlag.incomingDln.name);
+
+  bool isTransactionRequestEnabled() =>
+      _remoteConfig.getBool(FeatureFlag.transactionRequest.name);
 }
 
-enum FeatureFlag { outgoingDln, incomingDln }
+enum FeatureFlag { outgoingDln, incomingDln, transactionRequest }